+++ edfi.fif.api/src/graphql/modules/surveysummaryquestions.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import SurveySummaryQuestionsResolvers from '../resolvers/surveysummaryquestions.resolver';
 import SurveySummaryQuestionsService from '../services/surveysummaryquestions.service';
 import SurveySummaryQuestionsEntity from '../entities/survey/surveysummaryquestions.entity';
-import SurveySummaryAnsweresEntity from '../entities/survey/surveysummaryansweres.entity';
+import SurveySummaryAnsweresEntity from '../entities/survey/surveysummaryanswers.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([SurveySummaryQuestionsEntity, SurveySummaryAnsweresEntity])],
